rss_telegram_bot/requirements.txt
# Основные зависимости
python-telegram-bot[ext]>=20.0 # Библиотека для Telegram Bot API с расширениями
SQLAlchemy>=1.4 # ORM для работы с базой данных
feedparser>=6.0.11 # Для парсинга RSS/Atom лент (версия без cgi)
APScheduler>=3.9 # Для планирования задач (проверка лент)
httpx>=0.24.0 # HTTP клиент, используемый python-telegram-bot
httpcore>=0.17.0 # Низкоуровневая HTTP библиотека, используемая httpx

# Опционально, но рекомендуется для управления переменными окружения
python-dotenv>=0.19

# Для цветного вывода в консоль
rich>=13.0

# Зависимости для SQLAlchemy с SQLite (обычно включены в Python, но явно указать не помешает)
# sqlite3 - встроенный модуль

# Зависимости feedparser (обычно устанавливаются автоматически)
# sgmllib3k (для Python 3)

# Для работы с датами и временем (встроенные)
# datetime
# time

# Для работы с URL (встроенный)
# urllib

# Для логирования (встроенный)
# logging

# Для работы с HTML (встроенный)
# html
